Move some classes to private
